Introduce AiServices SPI
