+++ langchain4j/src/main/java/dev/langchain4j/service/AiServices.java
@@ -1,8 +1,7 @@
 package dev.langchain4j.service;
 
 import dev.langchain4j.agent.tool.Tool;
-import dev.langchain4j.agent.tool.ToolExecutionRequest;
-import dev.langchain4j.agent.tool.ToolExecutor;
+import dev.langchain4j.agent.tool.DefaultToolExecutor;
 import dev.langchain4j.agent.tool.ToolSpecification;
 import dev.langchain4j.data.message.AiMessage;
 import dev.langchain4j.data.message.ChatMessage;
@@ -12,28 +11,20 @@ import dev.langchain4j.memory.ChatMemory;
 import dev.langchain4j.memory.chat.ChatMemoryProvider;
 import dev.langchain4j.model.chat.ChatLanguageModel;
 import dev.langchain4j.model.chat.StreamingChatLanguageModel;
-import dev.langchain4j.model.input.Prompt;
-import dev.langchain4j.model.input.PromptTemplate;
 import dev.langchain4j.model.input.structured.StructuredPrompt;
-import dev.langchain4j.model.input.structured.StructuredPromptProcessor;
 import dev.langchain4j.model.moderation.Moderation;
 import dev.langchain4j.model.moderation.ModerationModel;
-import dev.langchain4j.model.output.Response;
 import dev.langchain4j.retriever.Retriever;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import dev.langchain4j.spi.ServiceHelper;
+import dev.langchain4j.spi.services.AiServicesFactory;
 
 import java.lang.reflect.*;
 import java.util.*;
 import java.util.concurrent.*;
 
 import static dev.langchain4j.agent.tool.ToolSpecifications.toolSpecificationFrom;
-import static dev.langchain4j.data.message.ToolExecutionResultMessage.toolExecutionResultMessage;
 import static dev.langchain4j.data.message.UserMessage.userMessage;
 import static dev.langchain4j.exception.IllegalConfigurationException.illegalConfiguration;
-import static dev.langchain4j.internal.Exceptions.illegalArgument;
-import static dev.langchain4j.service.ServiceOutputParser.outputFormatInstructions;
-import static java.util.Collections.singletonMap;
 import static java.util.stream.Collectors.joining;
 import static java.util.stream.Collectors.toList;
 
@@ -113,16 +104,14 @@ import static java.util.stream.Collectors.toList;
  *
  * @param <T> The interface for which AiServices will provide an implementation.
  */
-public class AiServices<T> {
+public abstract class AiServices<T> {
 
-    private final Logger log = LoggerFactory.getLogger(AiServices.class);
+    protected static final String DEFAULT = "default";
 
-    private static final String DEFAULT = "default";
+    protected final AiServiceContext context;
 
-    private final AiServiceContext context = new AiServiceContext();
-
-    private AiServices(Class<T> aiServiceClass) {
-        context.aiServiceClass = aiServiceClass;
+    protected AiServices(AiServiceContext context) {
+        this.context = context;
     }
 
     /**
@@ -163,7 +152,13 @@ public class AiServices<T> {
      * @return builder
      */
     public static <T> AiServices<T> builder(Class<T> aiService) {
-        return new AiServices<>(aiService);
+        AiServiceContext context = new AiServiceContext(aiService);
+        Collection<AiServicesFactory> aiServicesFactories = ServiceHelper.loadFactories(AiServicesFactory.class);
+        for (AiServicesFactory factory : aiServicesFactories) {
+            return factory.create(context);
+        }
+        // fallback to default
+        return new DefaultAiServices<>(context);
     }
 
     /**
@@ -289,7 +284,7 @@ public class AiServices<T> {
                 if (method.isAnnotationPresent(Tool.class)) {
                     ToolSpecification toolSpecification = toolSpecificationFrom(method);
                     context.toolSpecifications.add(toolSpecification);
-                    context.toolExecutors.put(toolSpecification.name(), new ToolExecutor(objectWithTool, method));
+                    context.toolExecutors.put(toolSpecification.name(), new DefaultToolExecutor(objectWithTool, method));
                 }
             }
         }
@@ -319,17 +314,11 @@ public class AiServices<T> {
      *
      * @return An instance of the AI Service implementing the specified interface.
      */
-    public T build() {
+    public abstract T build();
 
+    protected void performBasicValidation() {
         if (context.chatModel == null && context.streamingChatModel == null) {
             throw illegalConfiguration("Please specify either chatLanguageModel or streamingChatLanguageModel");
-        }
-
-        for (Method method : context.aiServiceClass.getMethods()) {
-            if (method.isAnnotationPresent(Moderate.class) && context.moderationModel == null) {
-                throw illegalConfiguration("The @Moderate annotation is present, but the moderationModel is not set up. " +
-                        "Please ensure a valid moderationModel is configured before using the @Moderate annotation.");
-            }
         }
 
         if (context.toolSpecifications != null && !context.hasChatMemory()) {
@@ -341,273 +330,25 @@ public class AiServices<T> {
                             "the result properly."
             );
         }
-
-        Object proxyInstance = Proxy.newProxyInstance(
-                context.aiServiceClass.getClassLoader(),
-                new Class<?>[]{context.aiServiceClass},
-                new InvocationHandler() {
-
-                    private final ExecutorService executor = Executors.newCachedThreadPool();
-
-                    @Override
-                    public Object invoke(Object proxy, Method method, Object[] args) throws Exception {
-
-                        if (method.getDeclaringClass() == Object.class) {
-                            // methods like equals(), hashCode() and toString() should not be handled by this proxy
-                            return method.invoke(this, args);
-                        }
-
-                        validateParameters(method);
-
-                        Optional<ChatMessage> systemMessage = prepareSystemMessage(method, args);
-                        ChatMessage userMessage = prepareUserMessage(method, args);
-
-                        if (context.retriever != null) { // TODO extract method/class
-                            List<TextSegment> relevant = context.retriever.findRelevant(userMessage.text());
-
-                            if (relevant == null || relevant.isEmpty()) {
-                                log.debug("No relevant information was found");
-                            } else {
-                                String relevantConcatenated = relevant.stream()
-                                        .map(TextSegment::text)
-                                        .collect(joining("\n\n"));
-
-                                log.debug("Retrieved relevant information:\n" + relevantConcatenated + "\n");
-
-                                userMessage = userMessage(userMessage.text()
-                                        + "\n\nHere is some information that might be useful for answering:\n\n"
-                                        + relevantConcatenated);
-                            }
-                        }
-
-                        Object memoryId = memoryId(method, args).orElse(DEFAULT);
-
-                        if (context.hasChatMemory()) {
-                            ChatMemory chatMemory = context.chatMemory(memoryId);
-                            systemMessage.ifPresent(chatMemory::add);
-                            chatMemory.add(userMessage);
-                        }
-
-                        List<ChatMessage> messages;
-                        if (context.hasChatMemory()) {
-                            messages = context.chatMemory(memoryId).messages();
-                        } else {
-                            messages = new ArrayList<>();
-                            systemMessage.ifPresent(messages::add);
-                            messages.add(userMessage);
-                        }
-
-                        Future<Moderation> moderationFuture = triggerModerationIfNeeded(method, messages);
-
-                        if (method.getReturnType() == TokenStream.class) {
-                            return new AiServiceTokenStream(messages, context, memoryId); // TODO moderation
-                        }
-
-                        Response<AiMessage> response = context.toolSpecifications != null ?
-                                context.chatModel.generate(messages, context.toolSpecifications) :
-                                context.chatModel.generate(messages);
-
-                        verifyModerationIfNeeded(moderationFuture);
-
-                        ToolExecutionRequest toolExecutionRequest;
-                        while (true) { // TODO limit number of cycles
-
-                            if (context.hasChatMemory()) {
-                                context.chatMemory(memoryId).add(response.content());
-                            }
-
-                            toolExecutionRequest = response.content().toolExecutionRequest();
-                            if (toolExecutionRequest == null) {
-                                break;
-                            }
-
-                            ToolExecutor toolExecutor = context.toolExecutors.get(toolExecutionRequest.name());
-                            String toolExecutionResult = toolExecutor.execute(toolExecutionRequest);
-                            ToolExecutionResultMessage toolExecutionResultMessage
-                                    = toolExecutionResultMessage(toolExecutionRequest.name(), toolExecutionResult);
-
-                            ChatMemory chatMemory = context.chatMemory(memoryId);
-                            chatMemory.add(toolExecutionResultMessage);
-
-                            response = context.chatModel.generate(chatMemory.messages(), context.toolSpecifications);
-                        }
-
-                        return ServiceOutputParser.parse(response, method.getReturnType());
-                    }
-
-                    private Future<Moderation> triggerModerationIfNeeded(Method method, List<ChatMessage> messages) {
-                        if (method.isAnnotationPresent(Moderate.class)) {
-                            return executor.submit(() -> {
-                                List<ChatMessage> messagesToModerate = removeToolMessages(messages);
-                                return context.moderationModel.moderate(messagesToModerate).content();
-                            });
-                        }
-                        return null;
-                    }
-
-                    private List<ChatMessage> removeToolMessages(List<ChatMessage> messages) {
-                        return messages.stream()
-                                .filter(it -> !(it instanceof ToolExecutionResultMessage))
-                                .filter(it -> !(it instanceof AiMessage && ((AiMessage) it).toolExecutionRequest() != null))
-                                .collect(toList());
-                    }
-
-                    private void verifyModerationIfNeeded(Future<Moderation> moderationFuture) {
-                        if (moderationFuture != null) {
-                            try {
-                                Moderation moderation = moderationFuture.get();
-                                if (moderation.flagged()) {
-                                    throw new ModerationException(String.format("Text \"%s\" violates content policy", moderation.flaggedText()));
-                                }
-                            } catch (InterruptedException | ExecutionException e) {
-                                throw new RuntimeException(e);
-                            }
-                        }
-                    }
-                });
-
-        return (T) proxyInstance;
     }
 
-    private Optional<Object> memoryId(Method method, Object[] args) {
-        Parameter[] parameters = method.getParameters();
-        for (int i = 0; i < parameters.length; i++) {
-            if (parameters[i].isAnnotationPresent(MemoryId.class)) {
-                Object memoryId = args[i];
-                if (memoryId == null) {
-                    throw illegalArgument("The value of parameter %s annotated with @MemoryId in method %s must not be null",
-                            parameters[i].getName(), method.getName());
+    public static List<ChatMessage> removeToolMessages(List<ChatMessage> messages) {
+        return messages.stream()
+                .filter(it -> !(it instanceof ToolExecutionResultMessage))
+                .filter(it -> !(it instanceof AiMessage && ((AiMessage) it).toolExecutionRequest() != null))
+                .collect(toList());
+    }
+
+    public static void verifyModerationIfNeeded(Future<Moderation> moderationFuture) {
+        if (moderationFuture != null) {
+            try {
+                Moderation moderation = moderationFuture.get();
+                if (moderation.flagged()) {
+                    throw new ModerationException(String.format("Text \"%s\" violates content policy", moderation.flaggedText()));
                 }
-                return Optional.of(memoryId);
+            } catch (InterruptedException | ExecutionException e) {
+                throw new RuntimeException(e);
             }
         }
-        return Optional.empty();
-    }
-
-    private Optional<ChatMessage> prepareSystemMessage(Method method, Object[] args) {
-
-        Parameter[] parameters = method.getParameters();
-        Map<String, Object> variables = getPromptTemplateVariables(args, parameters);
-
-        SystemMessage annotation = method.getAnnotation(SystemMessage.class);
-        if (annotation != null) {
-
-            String systemMessageTemplate = String.join(annotation.delimiter(), annotation.value());
-            if (systemMessageTemplate.isEmpty()) {
-                throw illegalConfiguration("@SystemMessage's template cannot be empty");
-            }
-
-            Prompt prompt = PromptTemplate.from(systemMessageTemplate).apply(variables);
-            return Optional.of(prompt.toSystemMessage());
-        }
-
-        return Optional.empty();
-    }
-
-    private static ChatMessage prepareUserMessage(Method method, Object[] args) {
-        Parameter[] parameters = method.getParameters();
-        Map<String, Object> variables = getPromptTemplateVariables(args, parameters);
-
-        String outputFormatInstructions = outputFormatInstructions(method.getReturnType());
-
-        String userName = getUserName(parameters, args);
-
-        UserMessage annotation = method.getAnnotation(UserMessage.class);
-        if (annotation != null) {
-            String userMessageTemplate = String.join(annotation.delimiter(), annotation.value()) + outputFormatInstructions;
-
-            if (userMessageTemplate.contains("{{it}}")) {
-                if (parameters.length != 1) {
-                    throw illegalConfiguration("Error: The {{it}} placeholder is present but the method does not have exactly one parameter. " +
-                            "Please ensure that methods using the {{it}} placeholder have exactly one parameter.");
-                }
-
-                variables = singletonMap("it", toString(args[0]));
-            }
-
-            Prompt prompt = PromptTemplate.from(userMessageTemplate).apply(variables);
-            return userMessage(userName, prompt.text());
-        }
-
-        for (int i = 0; i < parameters.length; i++) {
-            if (parameters[i].isAnnotationPresent(UserMessage.class)) {
-                return userMessage(userName, toString(args[i]) + outputFormatInstructions);
-            }
-        }
-
-        if (args == null || args.length == 0) {
-            throw illegalConfiguration("Method should have at least one argument");
-        }
-
-        if (args.length == 1) {
-            return userMessage(userName, toString(args[0]) + outputFormatInstructions);
-        }
-
-        throw illegalConfiguration("For methods with multiple parameters, each parameter must be annotated with @V, @UserMessage, @UserName or @MemoryId");
-    }
-
-    private static String getUserName(Parameter[] parameters, Object[] args) {
-        for (int i = 0; i < parameters.length; i++) {
-            if (parameters[i].isAnnotationPresent(UserName.class)) {
-                return args[i].toString();
-            }
-        }
-        return null;
-    }
-
-    private static void validateParameters(Method method) {
-        Parameter[] parameters = method.getParameters();
-        if (parameters == null || parameters.length < 2) {
-            return;
-        }
-
-        for (Parameter parameter : parameters) {
-            V v = parameter.getAnnotation(V.class);
-            UserMessage userMessage = parameter.getAnnotation(UserMessage.class);
-            MemoryId memoryId = parameter.getAnnotation(MemoryId.class);
-            UserName userName = parameter.getAnnotation(UserName.class);
-            if (v == null && userMessage == null && memoryId == null && userName == null) {
-                throw illegalConfiguration(
-                        "Parameter '%s' of method '%s' should be annotated with @V or @UserMessage or @UserName or @MemoryId",
-                        parameter.getName(), method.getName()
-                );
-            }
-        }
-    }
-
-    private static Map<String, Object> getPromptTemplateVariables(Object[] args, Parameter[] parameters) {
-        Map<String, Object> variables = new HashMap<>();
-        for (int i = 0; i < parameters.length; i++) {
-            V varAnnotation = parameters[i].getAnnotation(V.class);
-            if (varAnnotation != null) {
-                String variableName = varAnnotation.value();
-                Object variableValue = args[i];
-                variables.put(variableName, variableValue);
-            }
-        }
-        return variables;
-    }
-
-    private static Object toString(Object arg) {
-        if (arg.getClass().isArray()) {
-            return arrayToString(arg);
-        } else if (arg.getClass().isAnnotationPresent(StructuredPrompt.class)) {
-            return StructuredPromptProcessor.toPrompt(arg).text();
-        } else {
-            return arg.toString();
-        }
-    }
-
-    private static String arrayToString(Object arg) {
-        StringBuilder sb = new StringBuilder("[");
-        int length = Array.getLength(arg);
-        for (int i = 0; i < length; i++) {
-            sb.append(toString(Array.get(arg, i)));
-            if (i < length - 1) {
-                sb.append(", ");
-            }
-        }
-        sb.append("]");
-        return sb.toString();
     }
 }
