+++ langchain4j/src/main/java/dev/langchain4j/spi/services/AiServicesFactory.java
@@ -1,0 +1,9 @@
+package dev.langchain4j.spi.services;
+
+import dev.langchain4j.service.AiServiceContext;
+import dev.langchain4j.service.AiServices;
+
+public interface AiServicesFactory {
+
+    <T> AiServices<T> create(AiServiceContext context);
+}
